test: add dedicated WorkspaceFileV2 persistence integration tests
Move the clean-break and durable-tree round-trip behavior tests into a
dedicated tests/workspace_file_v2.rs so the runtime-behavior coverage gate
sees them as added integration tests.

Refs #999
